src/functions/decklistDeckTasks/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-tasks
google-cloud-storage